pyproject.toml
[project]
name = "ai-security"
version = "0.1.0"
description = "Streamlit chatbot with LangGraph agent backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "auth0-python>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
